+++ stable_baselines/__init__.py
@@ -8,4 +8,4 @@ from stable_baselines.ppo1 import PPO1
 from stable_baselines.ppo2 import PPO2
 from stable_baselines.trpo_mpi import TRPO
 
-__version__ = "1.0.8.rc3"
+__version__ = "2.0.0.rc0"